first finish on AS, setup RequestHandler + RequestListener + fix return type RequestNode

+++ weather-aggregator/src/main/java/com/weather/server/helper/RequestNode.java
@@ -7,16 +7,16 @@ import java.net.Socket;
 public class RequestNode {
     private final Socket socket;
     private final Request request;
-    private final long lamportClockValue;
+    private final int lamportClockValue;
 
     public RequestNode(Socket socket, Request request) {
         this.socket = socket;
         this.request = request;
         // Get the Lamport clock value from the request headers
-        this.lamportClockValue = Long.parseLong(request.getHeaders().get("Lamport-Clock"));
+        this.lamportClockValue = Integer.parseInt(request.getHeaders().get("Lamport-Clock"));
     }
 
     public Socket getSocket() { return socket; }
     public Request getRequest() { return request; }
-    public long getLamportClockValue() { return lamportClockValue; }
+    public int getLamportClockValue() { return lamportClockValue; }
 }
